Multiple updates
My tracking is getting worse through time, since this is really
important at the outset. For example, I'm still unsure about the
"proper" ways to install Spark and R, and additional packages for both.
I'll try to keep things more organized as I go through some of the
classes I purchased, and that should be good for the time being.